+++ dartstream/dartstream_backend/packages/frameworks/flutter_mobile/lib/api/ds_flutter_mobile_api.dart
@@ -1,0 +1,4 @@
+/* Used for API-related logic (if applicable), such as REST or GraphQL integrations, 
+SDK bindings, or framework-specific network utilities. */
+
+class DSFLutterMobileApiCore {}