+++ dartstream/dartstream_backend/packages/frameworks/flutter_mobile/lib/core/ds_flutter_mobile_core_export.dart
@@ -1,0 +1,8 @@
+library ds_flutter_mobile_core_export;
+
+// Export standard core module
+export 'ds_flutter_mobile_core.dart';
+export 'ds_flutter_mobile_auth_adapter.dart';
+
+
+// You can also include any shared utility functions or common interfaces here that might be used across multiple utility files.