+++ dartstream/dartstream_backend/packages/frameworks/flutter_mobile/lib/extensions/ds_flutter_mobile_extensions.dart
@@ -1,0 +1,3 @@
+/*Users wanting to extend the framework see the extensions/ folder as the dedicated space for their work.*/
+
+class DSFlutterMobileExtensions {}