+++ dartstream/dartstream_backend/packages/frameworks/flutter_mobile/lib/utilities/ds_flutter_mobile_utilities.dart
@@ -1,0 +1,4 @@
+/*Contains framework-specific helper functions, such as formatters, validators, 
+or commonly used methods. This folder helps modularize non-core logic. */
+
+class DSFlutterMobileUtilties {}