+++ dartstream/dartstream_backend/packages/frameworks/flutter_mobile/lib/utilities/ds_flutter_mobile_utilities_export.dart
@@ -1,0 +1,7 @@
+library ds_standard_utilities;
+
+// Export standard utilities module
+export 'ds_flutter_mobile_utilities.dart';
+
+
+// You can also include any shared utility functions or common interfaces here that might be used across multiple utility files.