+++ dartstream/dartstream_backend/packages/tooling/ds_tools/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.0.1-pre+3deprecated] - 2024-11-25
+
+### Deprecated
+- Marked the `ds_tools_cli` package as deprecated.
+- Added `deprecated.md` with detailed migration instructions.
+- Transitioned CLI functionality to the `ds_cli` in the `bin` folder of the DartStream framework.
+- Users are encouraged to migrate to the updated CLI tools provided in the main framework.
+
 ## [0.0.1-pre+2] - 2024-06-7
 
 ### Added